@@ -3,7 +3,6 @@ package blogspot.software_and_algorithms.stern_library.geometry;
 import java.awt.geom.Point2D;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Comparator;
 import java.util.List;
 
 import org.junit.Assert;
@@ -37,15 +36,12 @@ import org.junit.Test;
  */
 public class ClosestPointPairAlgorithmTest {
   private static Point2D[] sort(Point2D[] points) {
-    Arrays.sort(points, new Comparator<Point2D>() {
-      @Override
-      public int compare(Point2D o1, Point2D o2) {
-        double d = o1.getX() - o2.getX();
-        if (d == 0) {
-          d = o1.getY() - o2.getY();
-        }
-        return d < 0 ? -1 : d > 0 ? 1 : 0;
+    Arrays.sort(points, (o1, o2) -> {
+      double d = o1.getX() - o2.getX();
+      if (d == 0) {
+        d = o1.getY() - o2.getY();
       }
+      return d < 0 ? -1 : d > 0 ? 1 : 0;
     });
     return points;
   }
@@ -58,9 +54,9 @@ public class ClosestPointPairAlgorithmTest {
     List<Point2D> list = new ArrayList<Point2D>();
     list.add(p1);
     list.add(p2);
-    Assert.assertTrue(Arrays.equals(new Point2D[] { p1, p2 },
-                                    sort(new ClosestPointPairAlgorithm(list)
-                                        .execute())));
+    Assert.assertTrue(Arrays.equals(
+        new Point2D[] { p1, p2 },
+        sort(new ClosestPointPairAlgorithm(list).execute())));
 
   }
 
@@ -74,9 +70,9 @@ public class ClosestPointPairAlgorithmTest {
     list.add(p1);
     list.add(p2);
     list.add(p3);
-    Assert.assertTrue(Arrays.equals(new Point2D[] { p1, p3 },
-                                    sort(new ClosestPointPairAlgorithm(list)
-                                        .execute())));
+    Assert.assertTrue(Arrays.equals(
+        new Point2D[] { p1, p3 },
+        sort(new ClosestPointPairAlgorithm(list).execute())));
 
   }
 
@@ -91,9 +87,9 @@ public class ClosestPointPairAlgorithmTest {
     list.add(new Point2D.Double(.6, .5));
     list.add(new Point2D.Double(.7, .5));
     list.add(new Point2D.Double(.8, .5));
-    Assert.assertTrue(Arrays.equals(new Point2D[] { p1, p2 },
-                                    sort(new ClosestPointPairAlgorithm(list)
-                                        .execute())));
+    Assert.assertTrue(Arrays.equals(
+        new Point2D[] { p1, p2 },
+        sort(new ClosestPointPairAlgorithm(list).execute())));
   }
 
   @Test
@@ -107,8 +103,8 @@ public class ClosestPointPairAlgorithmTest {
     list.add(new Point2D.Double(.5, .2));
     list.add(new Point2D.Double(.4, .5));
     list.add(new Point2D.Double(.6, .5));
-    Assert.assertTrue(Arrays.equals(new Point2D[] { p1, p2 },
-                                    sort(new ClosestPointPairAlgorithm(list)
-                                        .execute())));
+    Assert.assertTrue(Arrays.equals(
+        new Point2D[] { p1, p2 },
+        sort(new ClosestPointPairAlgorithm(list).execute())));
   }
 }
